+++ docs/changes.md
@@ -5,6 +5,11 @@ title: Auth Change Log
 
 ## Version 3.1 {#v3-1}
 
+### v3.1.4 {#v3-1-4}
+
+* Improved performances by reducing call within `Illuminate\Container\Container`.
+* Reduces call to `Orchestra\Authorization\Keyword` on known keyword.
+
 ### v3.1.3 {#v3-1-3}
 
 * Authorization: